requirements.txt
asyncio
pyaudio
pyttsx3
python-dotenv
faster-whisper
openai
pygame
fastapi
uvicorn[standard]
torch
torchaudio
transformers
webrtcvad
pydub
ffmpeg